refactor(tests): split the two files closest to the line-limit cap
`line-limit` reads the whole tree, not the diff, so a file that has crept to
within a few lines of the 500-line hard limit fails whichever pull request next
touches it. The pull request that breaks is not the one that made the file long.
Four have been tumbled that way already.

Measured with the gate's own counting rule rather than an approximation of it —
`reusable-line-limit.yml` sets `hash_comment=0` for Rust, so `#[derive]` and
`#[cfg]` count as code, which my first pass wrongly stripped. On the real rule
thirteen files sit between 451 and 500. These are the two highest:

  internal/engine/lifecycle/tests.rs   487 -> 212, plus teardown_tests.rs 285
  internal/quadlet/tests/fields.rs     481 -> 365, plus fields_resources.rs 119

Both are test files with no inline module, which is why these two and not the
other eleven. `cargo llvm-cov` does not report on sibling test files — measured:
`lifecycle/tests.rs` and `quadlet/tests/fields.rs` are absent from the coverage
data, while `quadlet/warnings.rs` appears at 95.0% because its 271 lines of
inline tests are counted as covered lines of the file. So splitting these two
cannot move the coverage number, and splitting the other eleven by moving their
inline tests out would, against a threshold of 79 with only about three points
of headroom. That is a decision, not a refactor, so it is not made here.

Both cuts are by subject and both follow a precedent in the same directory.
`scale_tests.rs` and `drop_recheck_tests.rs` were already split off the lifecycle
suite, and `fields_logging.rs` off `fields.rs` for this exact reason (#1354).
What stays in `tests.rs` brings a project up; what moves takes one down. What
stays in `fields.rs` maps compose fields to Quadlet keys; what moves renders a
limit or a `deploy:` key.

`teardown_tests.rs` carries its own `engine_with` rather than importing one,
which is what `scale_tests.rs` does: three duplicated lines read better than a
helper exported out of a test module.

Two things clippy caught that I had not: `tests.rs` was left importing
`container_rm_path`, `ComposeFile` and `Service` after the tests using them
moved, and the `#[cfg(unix)]` guards had to be mirrored onto the new file's
imports and helper or they would be dead code on Windows, where the suite builds
with `-D warnings`. I could not cross-compile to check that here (no mingw), so
the guard structure mirrors the original file, which did build there.

1731 unit tests pass, the same count as before, so nothing was dropped in the move.
